+++ CHANGELOG.md
@@ -65,6 +65,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Made clouds bigger, more performant and prettier
 - Terrain meshing optimized further
 - Tree leaves no longer color blended
+- Actual character stats displayed in character window
 
 ### Removed